+++ crates/serialization/src/python/mod.rs
@@ -15,6 +15,7 @@
 
 //! Python bindings from [PyO3](https://pyo3.rs).
 
+#[cfg(feature = "arrow")]
 pub mod arrow;
 
 use pyo3::prelude::*;
@@ -26,46 +27,49 @@ use pyo3::prelude::*;
 /// Returns a `PyErr` if registering any module components fails.
 #[pymodule]
 pub fn serialization(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::get_arrow_schema_map,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::pyobjects_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_book_deltas_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_book_depth10_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_quotes_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_trades_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_bars_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_mark_prices_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_index_prices_to_arrow_record_batch_bytes,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::python::arrow::py_instrument_closes_to_arrow_record_batch_bytes,
-        m
-    )?)?;
+    #[cfg(feature = "arrow")]
+    {
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::get_arrow_schema_map,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::pyobjects_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_book_deltas_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_book_depth10_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_quotes_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_trades_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_bars_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_mark_prices_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_index_prices_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(
+            crate::python::arrow::py_instrument_closes_to_arrow_record_batch_bytes,
+            m
+        )?)?;
+    }
 
     Ok(())
 }